py-scipy: updated to 1.8.0
SciPy 1.8.0 is the culmination of 6 months of hard work. It contains
many new features, numerous bug-fixes, improved test coverage and better
documentation. There have been a number of deprecations and API changes
in this release, which are documented below. All users are encouraged to
upgrade to this release, as there are a large number of bug-fixes and
optimizations. Before upgrading, we recommend that users check that
their own code does not use deprecated SciPy functionality (to do so,
run your code with ``python -Wd`` and check for ``DeprecationWarning`` s).
Our development attention will now shift to bug-fix releases on the
1.8.x branch, and on adding new features on the master branch.

This release requires Python 3.8+ and NumPy 1.17.3 or greater.

For running on PyPy, PyPy3 6.0+ is required.


**************************
Highlights of this release
**************************

- A sparse array API has been added for early testing and feedback; this
  work is ongoing, and users should expect minor API refinements over
  the next few releases.
- The sparse SVD library PROPACK is now vendored with SciPy, and an interface
  is exposed via `scipy.sparse.svds` with ``solver='PROPACK'``. It is currently
  default-off due to potential issues on Windows that we aim to
  resolve in the next release, but can be optionally enabled at runtime for
  friendly testing with an environment variable setting of ``USE_PROPACK=1``.
- A new `scipy.stats.sampling` submodule that leverages the ``UNU.RAN`` C
  library to sample from arbitrary univariate non-uniform continuous and
  discrete distributions
- All namespaces that were private but happened to miss underscores in
  their names have been deprecated.