(math/R-googlesheets4) import R-googlesheets4-1.0.0
Interact with Google Sheets through the Sheets API v4
<https://developers.google.com/sheets/api>. "API" is an acronym for
"application programming interface"; the Sheets API allows users to
interact with Google Sheets programmatically, instead of via a web
browser. The "v4" refers to the fact that the Sheets API is currently
at version 4. This package can read and write both the metadata and
the cell data in a Sheet.