math/R-repurrrsive: import R-repurrrsive-1.0.0
Recursive lists in the form of R objects, 'JSON', and 'XML', for use
in teaching and examples. Examples include color palettes, Game of
Thrones characters, 'GitHub' users and repositories, music
collections, and entities from the Star Wars universe. Data from the
'gapminder' package is also included, as a simple data frame and in
nested and split forms.