opa: Update to 0.40.0
Changes:
0.40.0
======
This release contains a number of fixes and enhancements.

Metadata introspection
----------------------
The rich metadata added in the v0.38.0 release can now be introspected
from the policies themselves.

Both the rule's metadata can be accessed, via `rego.metadata.rule()`, and the
entire chain of metadata attached to the rule via the various scopes that different
metadata annotations can have, via `rego.metadata.chain()`.

Function mocking
----------------
It is now possible to mock functions in tests! Both built-in and
non-built-in functions can be mocked.

Assignments with `:=`
---------------------
Remaining restrictions around the use of `:=` in rules and functions
have been lifted (#4555).

In the wake of this, rules may now be "redeclared", i.e. you can use
`:=` for more than one rule body.

Experimental OCI support
------------------------
When configured to do so, OPA's bundle and discovery plugins will
retrieve bundles from any OCI registry.

Note that at this point, it's best considered a "feature preview".